refactor(#417 W-B5, C half): 107 deprecated forwarders retired, and the count in the doc was two different numbers
Every one had ZERO call sites, verified per name with `grep -rIw` and no
`--include` filter. Measured at the branch tip: 131 `NROS_DEPRECATED_MSG`
occurrences across 15 headers, of which 4 are the macro definition, 3 are
cbindgen doc prose, 1 is prose in `rcl_compat.h` and 13 are one banner line per
header -- leaving **107 forwarder declarations in 12 headers**, all unique
names.

The phase doc says "110 across 12 headers" and W-B5 says "43 names". Both are
wrong in the same direction as this campaign's other surveys: 43 is only the
stage-6 rcl/rclc subset, and the other 64 are phase-379-era renames
(`nros_param_*` -> `nros_parameter_*`, `try_recv` -> `take`, `send_reply` ->
`send_response`, `executor_stop` -> `cancel`). All 107 go, per the doc's own
"zero forwarders present" acceptance.

Also 4 typedef ALIASES in `parameter.h`. Those are the case W-B6's changelog has
to carry on its own: a renamed C type disappears with no deprecation warning at
all, because a typedef cannot carry one.

One retarget, not a deletion: `executor.h`'s issue-0338 macro
`nros_executor_register_timer` pointed at a forwarder, so it now names
`rclc_executor_add_timer`. It is the only member of that family whose target
moved -- the other nine name live exported symbols. That spelling no longer
warns on the way through, and the comment says so.

`nros_generated.h` was involved only as doc-comment prose. Five Rust doc
comments claimed the forwarders "survive as an `NROS_DEPRECATED_MSG` static
inline"; the Rust source was fixed and the header REGENERATED with
`cargo run -p nros-cbindgen-headers`, never hand-edited.

Collateral, all compiler-forced. Five deprecation probes deleted: they were
expected-failure compiles, and with the identifiers gone they would have kept
passing while asserting nothing -- the absorbing-verdict shape the lane's own
comments warn about. Four alias probes trimmed to their live half AND RENAMED,
so the filename stops describing a test that no longer exists. One real call
site the grep review had mis-filed as probe-only, caught by the compiler:
`nros_service_take_request` in `typed_service_handler.c`.

Independently broken and fixed while here: `#[cfg(kani)] mod verification` in
`nros-c/src/service.rs` called `nros_service_init` in eight harnesses -- a name
with no Rust definition, only ever a C `static inline`. `cargo kani` on nros-c
could not have compiled, and `just check` never sees `cfg(kani)`.

Verified: `just check c` (`All C checks passed!`), `just check cpp`, `just check
fast` (198 gates, 4 skipped on an environment precondition), `codegen_golden`
green, and a final per-name call-site sweep leaving 4 benign matches -- two
docstrings, one normalizer unit-test literal, one inert `#[ignore]`d stub.
